Cubeの処理をクラス化してモジュール分割
